use trait for parallel grid setup instead of code duplication
