+++ src/grid/parallel_grid.rs
@@ -1,12 +1,20 @@
 //! A parallel implementation of a grid
+use crate::element::reference_cell;
 use crate::grid::traits::{Grid, Topology};
+use crate::traits::grid::{Builder, ParallelBuilder, GridType};
 use crate::traits::types::{CellLocalIndexPair, Ownership, ReferenceCellType};
 use mpi::{
     request::WaitGuard,
     topology::Communicator,
-    traits::{Destination, Source},
+    traits::{Destination, Equivalence, Source, Buffer},
+};
+use rlst::{
+    rlst_dynamic_array2, Array, BaseArray,
+    RandomAccessMut, VectorContainer,
 };
 use std::collections::HashMap;
+
+type RlstMat<T> = Array<T, BaseArray<T, VectorContainer<T>, 2>, 2>;
 
 /// Grid local to a process
 pub struct LocalGrid<G: Grid> {
@@ -392,7 +400,7 @@ impl<'comm, C: Communicator, G: Grid> Topology for ParallelGrid<'comm, C, G> {
         self.local_grid.topology().face_flat_index_to_index(index)
     }
     fn cell_types(&self) -> &[ReferenceCellType] {
-        self.local_grid.topology().cell_types()
+        Topology::cell_types(self.local_grid.topology())
     }
 }
 
@@ -415,5 +423,341 @@ impl<'a, C: Communicator, G: Grid> Grid for ParallelGrid<'a, C, G> {
 
     fn is_serial(&self) -> bool {
         false
+    }
+}
+
+// TODO: pub(crate)
+/// Internal trait for building parallel grids
+pub trait ParallelGridBuilder
+{
+    /// The serial grid type used on each process
+    type G: Grid;
+
+    /// The id of each point
+    fn point_indices_to_ids(&self) -> &[usize];
+
+    /// The coordinates of each point
+    fn points(&self) -> &[<<Self as ParallelGridBuilder>::G as GridType>::T];
+
+    /// The id of each cell
+    fn cell_indices_to_ids(&self) -> &[usize];
+
+    /// The point of a cell
+    fn cell_points(&self, index: usize) -> &[usize];
+
+    /// The vertices of a cell
+    fn cell_vertices(&self, index: usize) -> &[usize];
+
+    /// The cell type of a cell
+    fn cell_type(&self, index: usize) -> ReferenceCellType;
+
+    #[allow(clippy::too_many_arguments)]
+    /// Create a serial grid on one process
+    fn create_serial_grid(
+        &self,
+        points: RlstMat<<<Self as ParallelGridBuilder>::G as GridType>::T>,
+        cells: &[usize],
+        point_indices_to_ids: Vec<usize>,
+        point_ids_to_indices: HashMap<usize, usize>,
+        cell_indices_to_ids: Vec<usize>,
+        cell_ids_to_indices: HashMap<usize, usize>,
+        edge_ids: HashMap<[usize; 2], usize>,
+    ) -> Self::G;
+
+    #[allow(clippy::too_many_arguments)]
+    /// Internal function to create a parallel grid
+    fn create_internal<'a, C: Communicator>(
+        &self,
+        comm: &'a C,
+        points: &[<<Self as ParallelGridBuilder>::G as GridType>::T],
+        point_ids: &[usize],
+        cells: &[usize],
+        cell_owners: &[usize],
+        cell_ids: &[usize],
+        vertex_owners: &[usize],
+        vertex_ids: &[usize],
+        edges: &[usize],
+        edge_owners: &[usize],
+        edge_ids: &[usize],
+    ) -> ParallelGrid<'a, C, Self::G> {
+        let npts = point_ids.len();
+
+        let mut coordinates = rlst_dynamic_array2!(<<Self as ParallelGridBuilder>::G as GridType>::T, [npts, 3]);
+        for i in 0..npts {
+            for j in 0..3 {
+                *coordinates.get_mut([i, j]).unwrap() = points[i * 3 + j];
+            }
+        }
+
+        let mut point_ids_to_indices = HashMap::new();
+        for (index, id) in point_ids.iter().enumerate() {
+            point_ids_to_indices.insert(*id, index);
+        }
+        let mut cell_ids_to_indices = HashMap::new();
+        for (index, id) in cell_ids.iter().enumerate() {
+            cell_ids_to_indices.insert(*id, index);
+        }
+
+        let mut edge_id_map = HashMap::new();
+        for (n, id) in edge_ids.iter().enumerate() {
+            edge_id_map.insert([edges[2 * n], edges[2 * n + 1]], *id);
+        }
+
+        let serial_grid = self.create_serial_grid(
+            coordinates,
+            cells,
+            point_ids.to_vec(),
+            point_ids_to_indices,
+            cell_ids.to_vec(),
+            cell_ids_to_indices,
+            edge_id_map,
+        );
+
+        let mut vertex_owner_map = HashMap::new();
+        for (id, owner) in vertex_ids.iter().zip(vertex_owners) {
+            vertex_owner_map.insert(*id, *owner);
+        }
+        let mut edge_owner_map = HashMap::new();
+        for (id, owner) in edge_ids.iter().zip(edge_owners) {
+            edge_owner_map.insert(*id, *owner);
+        }
+        let mut cell_owner_map = HashMap::new();
+        for (id, owner) in cell_ids.iter().zip(cell_owners) {
+            cell_owner_map.insert(*id, *owner);
+        }
+
+        ParallelGrid::new(
+            comm,
+            serial_grid,
+            vertex_owner_map,
+            edge_owner_map,
+            cell_owner_map,
+        )
+    }
+}
+
+impl<
+        const GDIM: usize,
+        B: ParallelGridBuilder + Builder<GDIM>,
+    > ParallelBuilder<GDIM> for B
+where Vec<<<B as ParallelGridBuilder>::G as GridType>::T>: Buffer,
+    <<B as ParallelGridBuilder>::G as GridType>::T: Equivalence
+{
+    type ParallelGridType<'a, C: Communicator + 'a> = ParallelGrid<'a, C, B::G>;
+
+    fn create_parallel_grid<'a, C: Communicator>(
+        self,
+        comm: &'a C,
+        cell_owners: &HashMap<usize, usize>,
+    ) -> Self::ParallelGridType<'a, C> {
+        let rank = comm.rank() as usize;
+        let size = comm.size() as usize;
+
+        let npts = self.point_indices_to_ids().len();
+        let ncells = self.cell_indices_to_ids().len();
+
+        // data used in computation
+        let mut vertex_owners = vec![(-1, 0); npts];
+        let mut vertex_counts = vec![0; size];
+        let mut cell_indices_per_proc = vec![vec![]; size];
+        let mut vertex_indices_per_proc = vec![vec![]; size];
+        let mut edge_owners = HashMap::new();
+        let mut edge_ids = HashMap::new();
+        let mut edge_counts = vec![0; size];
+        let mut edges_included_per_proc = vec![vec![]; size];
+        let mut edge_id = 0;
+
+        // data to send to other processes
+        let mut points_per_proc = vec![vec![]; size];
+        let mut point_ids_per_proc = vec![vec![]; size];
+        let mut cells_per_proc = vec![vec![]; size];
+        let mut cell_owners_per_proc = vec![vec![]; size];
+        let mut cell_ids_per_proc = vec![vec![]; size];
+        let mut vertex_owners_per_proc = vec![vec![]; size];
+        let mut edges_per_proc = vec![vec![]; size];
+        let mut edge_owners_per_proc = vec![vec![]; size];
+        let mut edge_ids_per_proc = vec![vec![]; size];
+
+        for (index, id) in self.cell_indices_to_ids().iter().enumerate() {
+            let owner = cell_owners[id];
+            // TODO: only assign owners to the first 3 or 4 vertices
+            for v in self.cell_points(index) {
+                if vertex_owners[*v].0 == -1 {
+                    vertex_owners[*v] = (owner as i32, vertex_counts[owner]);
+                }
+                if !vertex_indices_per_proc[owner].contains(v) {
+                    vertex_indices_per_proc[owner].push(*v);
+                    vertex_owners_per_proc[owner].push(vertex_owners[*v].0 as usize);
+                    for i in 0..GDIM {
+                        points_per_proc[owner].push(self.points()[v * GDIM + i])
+                    }
+                    point_ids_per_proc[owner].push(self.point_indices_to_ids()[*v]);
+                    vertex_counts[owner] += 1;
+                }
+            }
+        }
+
+        for (index, id) in self.cell_indices_to_ids().iter().enumerate() {
+            let ref_conn = &reference_cell::connectivity(self.cell_type(index))[1];
+            let owner = cell_owners[id];
+            for e in ref_conn {
+                let cell = self.cell_vertices(index);
+                let mut v0 = cell[e[0][0]];
+                let mut v1 = cell[e[0][1]];
+                if v0 > v1 {
+                    std::mem::swap(&mut v0, &mut v1);
+                }
+                if edge_owners.get_mut(&(v0, v1)).is_none() {
+                    edge_owners.insert((v0, v1), (owner, edge_counts[owner]));
+                    edge_ids.insert((v0, v1), edge_id);
+                    edge_id += 1;
+                    edges_included_per_proc[owner].push((v0, v1));
+                    edges_per_proc[owner].push(v0);
+                    edges_per_proc[owner].push(v1);
+                    edge_owners_per_proc[owner].push(edge_owners[&(v0, v1)].0);
+                    edge_ids_per_proc[owner].push(edge_ids[&(v0, v1)]);
+                    edge_counts[owner] += 1;
+                }
+            }
+        }
+
+        for index in 0..ncells {
+            for p in 0..size {
+                for v in self.cell_points(index) {
+                    if vertex_indices_per_proc[p].contains(v) {
+                        cell_indices_per_proc[p].push(index);
+                        break;
+                    }
+                }
+            }
+        }
+
+        for p in 0..size {
+            for index in &cell_indices_per_proc[p] {
+                let id = self.cell_indices_to_ids()[*index];
+                // TODO: only assign owners to the first 3 or 4 vertices
+                for v in self.cell_points(*index) {
+                    if !vertex_indices_per_proc[p].contains(v) {
+                        vertex_indices_per_proc[p].push(*v);
+                        vertex_owners_per_proc[p].push(vertex_owners[*v].0 as usize);
+                        for i in 0..GDIM {
+                            points_per_proc[p].push(self.points()[v * GDIM + i]);
+                        }
+                        point_ids_per_proc[p].push(self.point_indices_to_ids()[*v]);
+                    }
+                    cells_per_proc[p].push(
+                        vertex_indices_per_proc[p]
+                            .iter()
+                            .position(|&r| r == *v)
+                            .unwrap(),
+                    );
+                }
+                let ref_conn = &reference_cell::connectivity(self.cell_type(*index))[1];
+
+                for e in ref_conn {
+                    let cell = self.cell_vertices(*index);
+                    let mut v0 = cell[e[0][0]];
+                    let mut v1 = cell[e[0][1]];
+                    if v0 > v1 {
+                        std::mem::swap(&mut v0, &mut v1);
+                    }
+                    if !edges_included_per_proc[p].contains(&(v0, v1)) {
+                        edges_included_per_proc[p].push((v0, v1));
+                        edges_per_proc[p].push(v0);
+                        edges_per_proc[p].push(v1);
+                        edge_owners_per_proc[p].push(edge_owners[&(v0, v1)].0);
+                        edge_ids_per_proc[p].push(edge_ids[&(v0, v1)]);
+                    }
+                }
+
+                cell_ids_per_proc[p].push(id);
+                cell_owners_per_proc[p].push(cell_owners[&id]);
+            }
+        }
+
+        mpi::request::scope(|scope| {
+            for p in 1..size {
+                let _ = WaitGuard::from(
+                    comm.process_at_rank(p as i32)
+                        .immediate_send(scope, &points_per_proc[p]),
+                );
+                let _ = WaitGuard::from(
+                    comm.process_at_rank(p as i32)
+                        .immediate_send(scope, &point_ids_per_proc[p]),
+                );
+                let _ = WaitGuard::from(
+                    comm.process_at_rank(p as i32)
+                        .immediate_send(scope, &cells_per_proc[p]),
+                );
+                let _ = WaitGuard::from(
+                    comm.process_at_rank(p as i32)
+                        .immediate_send(scope, &cell_owners_per_proc[p]),
+                );
+                let _ = WaitGuard::from(
+                    comm.process_at_rank(p as i32)
+                        .immediate_send(scope, &cell_ids_per_proc[p]),
+                );
+                let _ = WaitGuard::from(
+                    comm.process_at_rank(p as i32)
+                        .immediate_send(scope, &vertex_owners_per_proc[p]),
+                );
+                let _ = WaitGuard::from(
+                    comm.process_at_rank(p as i32)
+                        .immediate_send(scope, &edges_per_proc[p]),
+                );
+                let _ = WaitGuard::from(
+                    comm.process_at_rank(p as i32)
+                        .immediate_send(scope, &edge_owners_per_proc[p]),
+                );
+                let _ = WaitGuard::from(
+                    comm.process_at_rank(p as i32)
+                        .immediate_send(scope, &edge_ids_per_proc[p]),
+                );
+            }
+        });
+        self.create_internal(
+            comm,
+            &points_per_proc[rank],
+            &point_ids_per_proc[rank],
+            &cells_per_proc[rank],
+            &cell_owners_per_proc[rank],
+            &cell_ids_per_proc[rank],
+            &vertex_owners_per_proc[rank],
+            &point_ids_per_proc[rank],
+            &edges_per_proc[rank],
+            &edge_owners_per_proc[rank],
+            &edge_ids_per_proc[rank],
+        )
+    }
+
+    fn receive_parallel_grid<C: Communicator>(
+        self,
+        comm: &C,
+        root_rank: usize,
+    ) -> ParallelGrid<'_, C, B::G> {
+        let root_process = comm.process_at_rank(root_rank as i32);
+
+        let (points, _status) = root_process.receive_vec::<<<Self as ParallelGridBuilder>::G as GridType>::T>();
+        let (point_ids, _status) = root_process.receive_vec::<usize>();
+        let (cells, _status) = root_process.receive_vec::<usize>();
+        let (cell_owners, _status) = root_process.receive_vec::<usize>();
+        let (cell_ids, _status) = root_process.receive_vec::<usize>();
+        let (vertex_owners, _status) = root_process.receive_vec::<usize>();
+        let (edges, _status) = root_process.receive_vec::<usize>();
+        let (edge_owners, _status) = root_process.receive_vec::<usize>();
+        let (edge_ids, _status) = root_process.receive_vec::<usize>();
+        self.create_internal(
+            comm,
+            &points,
+            &point_ids,
+            &cells,
+            &cell_owners,
+            &cell_ids,
+            &vertex_owners,
+            &point_ids,
+            &edges,
+            &edge_owners,
+            &edge_ids,
+        )
     }
 }
